requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
cloudinary==1.43.0
colorama==0.4.6
Flask==3.1.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
pyodbc==5.2.0
six==1.17.0
SQLAlchemy==2.0.39
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
gunicorn==23.0.0

